apps/hub-api/pyproject.toml
[project]
name = "alice-hub-api"
version = "0.1.0"
description = "Alice Home OS API runtime"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13,<2.0",
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "email-validator>=2.1,<3.0",
  "paho-mqtt>=2.1,<3.0",
  "pydantic-settings>=2.2,<3.0",
  "PyJWT>=2.8,<3.0",
  "sqlalchemy>=2.0,<3.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<1.0",
  "pytest>=8.0,<9.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
